Create filenames in ascending sort order
Using the hyphen "-" as the separator for filenames with collision has
makes these filenames sort earlier than the original file. It is
useful to be able to sort the generated files by their name and assume
that the last filename in the sorted list represents the latest sent
email.

Using an underscore solves this issue since it sorts later than a
period, e.g.

  >>> sorted(['2013-07-22-131955.eml', '2013-07-22-131955_1.eml', '2013-07-22-131955-1.eml'])
  ['2013-07-22-131955-1.eml', '2013-07-22-131955.eml', '2013-07-22-131955_1.eml']