Add args and kwargs in DebuggingServer init method
There are a couple more parameters other than the ones the DebuggingServer is
taking right now. We are ignoring them, but still, smtpd is passing them.

To accept them without questioning, I added the *args and **kwargs in the
constructor definition.

A more reliable, but less flexible solution could be to put there the named
arguments of the SMTPServer with the default values that are shown in the
documentation here:

    https://docs.python.org/3/library/smtpd.html#smtpd.SMTPServer

But it's much easier to be unaligned with the python library (for example the
decode_data argument will have False as default from 3.6).